requirements.txt
Flask==3.0.2
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0